Produce bucketspaces config for entire content cluster.
