+++ config-model/src/main/java/com/yahoo/vespa/model/content/cluster/ContentCluster.java
@@ -7,12 +7,11 @@ import com.yahoo.config.model.ConfigModelContext;
 import com.yahoo.config.model.producer.AbstractConfigProducerRoot;
 import com.yahoo.config.provision.ClusterSpec;
 import com.yahoo.config.provision.Environment;
-import com.yahoo.config.provision.RegionName;
-import com.yahoo.config.provision.SystemName;
 import com.yahoo.config.provision.Zone;
 import com.yahoo.vespa.config.content.MessagetyperouteselectorpolicyConfig;
 import com.yahoo.vespa.config.content.FleetcontrollerConfig;
 import com.yahoo.vespa.config.content.StorDistributionConfig;
+import com.yahoo.vespa.config.content.core.BucketspacesConfig;
 import com.yahoo.vespa.config.content.core.StorDistributormanagerConfig;
 import com.yahoo.documentmodel.NewDocumentType;
 import com.yahoo.documentapi.messagebus.protocol.DocumentProtocol;
@@ -59,7 +58,8 @@ public class ContentCluster extends AbstractConfigProducer implements StorDistri
         StorDistributormanagerConfig.Producer,
         FleetcontrollerConfig.Producer,
         MetricsmanagerConfig.Producer,
-        MessagetyperouteselectorpolicyConfig.Producer {
+        MessagetyperouteselectorpolicyConfig.Producer,
+        BucketspacesConfig.Producer {
 
     // TODO: Make private
     private String documentSelection;
@@ -693,5 +693,19 @@ public class ContentCluster extends AbstractConfigProducer implements StorDistri
             }
         }
 
+    }
+
+    private static final String DEFAULT_BUCKET_SPACE = "default";
+    private static final String GLOBAL_BUCKET_SPACE = "global";
+
+    @Override
+    public void getConfig(BucketspacesConfig.Builder builder) {
+        for (NewDocumentType docType : getDocumentDefinitions().values()) {
+            BucketspacesConfig.Documenttype.Builder docTypeBuilder = new BucketspacesConfig.Documenttype.Builder();
+            docTypeBuilder.name(docType.getName());
+            String bucketSpace = (isGloballyDistributed(docType) ? GLOBAL_BUCKET_SPACE : DEFAULT_BUCKET_SPACE);
+            docTypeBuilder.bucketspace(bucketSpace);
+            builder.documenttype(docTypeBuilder);
+        }
     }
 }
